requirements.txt
selenium
python-dotenv
bs4
yt-dlp
pytest
seleniumbase
